Update RNAseq1_unix.Rmd
Minor changes in text/typos removal, added "grep" as text editing option, changed S5_cds.fasta to S5_cds_output.fasta to make it clearer to students that this is a new file. wc -cwl was changed to wc -lwc because, at least on my machine, 'wc -wlc', 'wc -wcl' and other combinations all give as output l,w and c (in this order). Put ".gbk" instead of ".gds" and other extensions which were not in the folder. Added a few things about Python.